backend/requirements.txt
aniso8601==9.0.1
click==8.0.1
colorama==0.4.4
Flask==2.0.2
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyMySQL==1.0.2
pytz==2021.3
redis==3.5.3
six==1.16.0
SQLAlchemy==1.3.19
Werkzeug==2.0.2
PyJWT~=2.3.0
flask-script==2.0.6
flask-migrate==3.1.0
flask-limiter==2.1.3
requests==2.27.1
flask-cors==3.0.10
